+++ CHANGELOG.md
@@ -4,5 +4,6 @@ MSC changelog
 ### 1.0.8
 
 - Added default health-check endpoint
+- Added Spring integration test configuration to BaseTest class
 
 ### 1.0.7